Initial commit in new decoupled repository Version 0.1
The current state is a development state of the interface (iCub<->ANNarchy).
Programming language: C++11; Python interface with Cython

The interface has a modular structure:

	Interface_iCub: main part of the interface
			conclude the seperate modules

	Joint_Reader: joint reader module, capsulating the joint readout mechanisms

	Joint_Writer: joint writer module, capsulating the joint moving mechanisms

	Visual_Reader: visual reader module, capsulating the readout of the robot cameras, only monocular in this version

	Skin_Reader: tactile reader module, capsulating the readout of the tactile data from the artificial skin

Each module consists of a .hpp header file and a .cpp source file, which are seperated in the include and src folders.

INI_Reader in include: Module used to handle the .ini files; not self-written

iCubCPP.pyx: Wrapping file for the Cython connection to Python

joint_limits.ini: .ini file containing the joint limits for the iCub